add some checks before commands. created new scripts!

+++ create-symbolic-links.sh
@@ -1,13 +1,21 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
+if [ ! -e ~/.vimrc ]
+then
 echo "creating symbolic link for .vimrc"
 ln -s ~/dotfiles/vimrc ~/.vimrc
+fi
 
+if [ ! -e ~/.tmux.conf ]
+then
 echo "creating symbolic link for .tmux.conf"
 ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
+fi
 
-echo "creating symbolic link for .gitignore_global"
+if [ ! -e ~/.gitignore ]
+then
+echo "creating symbolic link for .gitignore"
 ln -s ~/dotfiles/gitignore ~/.gitignore
-git config --global core.excludesfile ~/.gitignore
+fi
 
 ls -al ~/.tmux.conf ~/.vimrc ~/.gitignore
